gosenProject/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.2
django-braces==1.13.0
django-cors-headers==3.0.2
django-rest-passwordreset==1.0.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
kombu==4.6.3
Pillow==6.0.0
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
django-rest-swagger==2.2.0